Tidy: createIcon moved to utils
